Add `url` to `DocumentObjectType` (#95)
* Add simple test for existing document schema

* Add test for fileHash and fileSize GraphQL fields

* Add assertions to test setup for document test

* Add test for image query

Image queries before have been tested through some page tests, but not
directly.

The new test is only minimal but could be extended for more image fields
that can be queried.

The test includes the `src` filed on the image query. This test serves as
an example to add the same test for the documents next and then implement
the filed on the document similar to how it exists on the image already.

* Change document test names for clarity

* Add failing test for documents.src field

The test is failing because the field does not exist on the endpoint yet.
Implementing the field and passing the test is the next step

* Duplicate src query test for different document serve methods

The URL in src will differ depending on the document serve method used.

The tests are now sensitive to this different and expect different values.
To make the default method explicit, the corresponding option has been added.

* Add src field to DocumentObjectType

The method to resolve the `src` field is basically the same as for the BaseImageObjectType.

* Blackify

* Change field name from src to url in test

This change was requested by @zerolabs in
https://github.com/GrappleGQL/wagtail-grapple/pull/96\#issuecomment-674950403
and
https://github.com/GrappleGQL/wagtail-grapple/issues/94\#issuecomment-674955038

* Update field name from src to url